refactor: Modal 디렉터리 수정
