+++ src/components/Post/PostModal.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import Modal from '../../shared/Modal/Modal.tsx'
+import Modal from '../../shared/Modal.tsx'
 import PostImage from './PostImage.tsx'
 import PostHeader from './PostHeader.tsx'
 import ReactSection from './ReactSection.tsx'
